Upgrade to datafusion 47 (#3016)
Getting ready for the [datafusion 47
release](https://github.com/apache/datafusion/issues/15072).

## Current issues
- [x] https://github.com/apache/datafusion/issues/15072
- [x] duckdb-rs depends on arrow 54, opened a PR to fix -
https://github.com/duckdb/duckdb-rs/pull/496.
- [ ] object-store 0.12 has a regression on Azure, not sure what's the
priority here but shouldn't be too hard to find the root cause if we
care. https://github.com/apache/arrow-rs-object-store/issues/320

+++ vortex-datafusion/src/persistent/opener.rs
@@ -144,7 +144,7 @@ fn apply_byte_range(
         let row_range = byte_range_to_row_range(
             byte_range.start as u64..byte_range.end as u64,
             row_count,
-            file_meta.object_meta.size as u64,
+            file_meta.object_meta.size,
         );
 
         scan_builder.with_row_range(row_range)
